Turn SwiftUI radio button into its own component
